+++ .github/scripts/docker-test.sh
@@ -1,6 +1,4 @@
-#!/bin/bash
-dotenv=$1
-
+#!/bin/sh
 cd /app
 # Installing Requirements
 echo "\033[33mTest 1: Installing requirements...\033[0m"
@@ -12,9 +10,13 @@ if [ $? -ne 0 ]; then
 fi
 echo "\033[32mTest 1: Completed\033[0m"
 
+# Creating .env file
+echo "TG_BOT_TOKEN='$TG_BOT_TOKEN'" > .env
+echo "PORT='$PORT'" >> .env
+
 # Check if script is running properly
 echo "\033[33mTest 2: Checking if script is running properly...\033[0m"
-echo $dotenv > .env
+
 if [ $? -ne 0 ]; then
     echo "\033[31mFailed to create .env file\033[0m"
     exit 1